+++ test/runtests.jl
@@ -18,5 +18,6 @@ using CairoMakie
     include("test_interventions.jl")
     include("test_dagplot.jl")
     include("test_visual_grammar.jl")
+    include("test_smart.jl")
     include("test_aqua.jl")
 end